Combine TS with Wizards (since all wizards now).
